consolidated abstract case and drivers for container runtime system
