--- beeflow/common/crt/Readme.md
@@ -1,1 +1,0 @@
-crt is Container Runtime
